dom/media/tests/identity/test_setIdentityProviderWithErrors.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="../mochitest/head.js"></script>
  <script type="application/javascript" src="../mochitest/pc.js"></script>
  <script type="application/javascript" src="../mochitest/templates.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    title: "Identity Provider returning errors is handled correctly"
  });

var test;
runTest(function () {
  test = new PeerConnectionTest();
  test.setMediaConstraints([{audio: true}], [{audio: true}]);
  // first example generates an error
  test.setIdentityProvider(test.pcLocal, 'example.com', 'idp.html#error', 'nobody');
  // first doesn't even get a ready message from the IdP - results in a timeout
  test.setIdentityProvider(test.pcRemote, 'example.com', 'idp.html#error:ready', 'nobody');
  test.chain.append([
  [
    "PEER_IDENTITY_IS_EMPTY",
    function(test) {
      ok(!test.pcLocal._pc.peerIdentity, "local peerIdentity is not set");
      ok(!test.pcRemote._pc.peerIdentity, "remote peerIdentity is not set");
      test.next();
    }
  ],
  [
    "OFFERS_AND_ANSWERS_DONT_INCLUDE_IDENTITY",
    function(test) {
      ok(!test.pcLocal._last_offer.sdp.contains("a=identity"), "a=identity not contained in the offer SDP");
      ok(!test.pcRemote._last_answer.sdp.contains("a=identity"), "a=identity not contained in the answer SDP");
      test.next();
    }
  ],
  ]);
  test.run();
});

</script>
</pre>
</body>
</html>
